replace conditional with polymorphism
